+++ modules/motions/Rotation.h
@@ -17,7 +17,7 @@ class Rotation : public Motion {
    * @param _speed       指定する速度（mm/秒）
    * @param _isClockwise 回頭方向 true:時計回り, false:反時計回り
    */
-  Rotation(Robot& _robot, int _speed, bool _isClockwise);
+  Rotation(Robot& _robot, double _speed, bool _isClockwise);
 
   /**
    * @brief 回頭する
@@ -45,7 +45,7 @@ class Rotation : public Motion {
   virtual bool isMetContinuationCondition() = 0;
 
  protected:
-  int speed;         // 指定する速度（mm/秒）
+  double speed;      // 指定する速度（mm/秒）
   bool isClockwise;  // 回頭方向 true:時計回り, false:反時計回り
   int leftSign;      // leftSign  = +1 のとき左モーターは順回転、 -1 のとき逆回転
   int rightSign;     // rightSign = +1 のとき右モーターは順回転、 -1 のとき逆回転
